+++ src/com/bean/Department.java
@@ -8,14 +8,14 @@ import javax.persistence.Table;
 @Table(name = "department")
 public class Department {
 	@Id
-	private int department_id;
+	private long department_id;
 	private String department_nm;
 	
-	public int getDepartment_id()
+	public long getDepartment_id()
 	{
 		return department_id;
 	}
-	public void setDepartment_id(int department_id)
+	public void setDepartment_id(long department_id)
 	{
 		this.department_id = department_id;
 	}
